requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
SQLAlchemy
psycopg2-binary
langchain
langchain-community
langchain-google-genai
google-api-core
langchain-classic
chromadb
finnhub-python
beautifulsoup4
html2text
numpy
pandas
tabulate
scikit-learn
redis
langgraph